# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
    description: "This is the REST 'contract' for the Order System"
    version: 1.0.0
    title: "(20F) Group Project - Order System"
    contact:
        name: "(20F) CST 8277"
        email: "normanm@algonquincollege.com"
        url: "https://brightspace.algonquincollege.com/d2l/home/279461"
servers:
    -
        url: "http://localhost:{port}/{contextRoot}/{basePath}"
        variables:
            port: {
                default: "9090"
            }
            contextRoot: {
                default: "rest-orderSystem"
            }
            basePath: {
                default: "api/v1"
            }
security:
    -
        basicAuth: []
components:
    securitySchemes:
        basicAuth:
            type: http
            scheme: basic
...
    responses:
...
        UnauthorizedAdminError:
            description: "request requires HTTP authentication for ADMIN_ROLE"
...
    schemas:
        PojoBase:
            type: object
            required: [id, version]
            properties: {
                id: {
                    description: "id of entity",
                    type: integer
                },
                version: {
                    description: "version (OptLock) of entity",
                    type: integer
                },
                createdDate: {
                    description: "creation timestamp of entity",
                    type: string,
                    format: date-time
                    },
                updatedDate: {
                    description: "last-updatd timestamp of entity",
                    type: string,
                    format: date-time
                }
            }
...
        Product:
            description: "(20F) Group Project - Product entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                description: {type: string, description: "Description of product", example: John},
                serialNo: {type: string, description: "serialNo of product", example: Doe},
                store: {$ref: '#/components/schemas/StoresList'}
            }
        Store:
            description: "(20F) Group Project - Store entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                storeName: {type: string, description: "Name of store", example: shopers},
                product: {$ref: '#/components/schemas/ProductsList'}
            }
...
        ProductsList:
            description: "list of products"
            type: array
            items: {
                $ref: "#/components/schemas/Product"
            }
        StoresList:
            description: "list of stores"
            type: array
            items: {
                $ref: "#/components/schemas/Store"
            }
...
paths:
...
    /store:
        get:
            summary: "{All} Get all stores"
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Store"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedAdminError"
                }
            }

        post:
            summary: "{admin} Post a store"
            requestBody: {
                description: "new store to be added to Db",
                required: true,
                content: {
                    application/json: {
                        schema: {
                            $ref: "#/components/schemas/Store"
                        }
                    }
                }
            }
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Store"
                            }
                        }
                    }
                }
            }
...